build: make generating make targets more systematic
This addresses a suggestion from a review comment in another PR.

The goal of this change is to make generating make targets more systematic.

Firstly, generating targets are made available under a more systematic naming
 following the pattern of "gen.foo".
more concretely, the following targets are introduced as aliases of the corresponding existing
 targets:

* gen.codegen
* gen.rbac
* gen.crds
* gen.docs
* gen.crd-docs

additionally, a  new "generate" make target is added that invokes all generating targets.

Co-authored-by: Blaine Gardner <b.blaine.gardner@gmail.com>
Signed-off-by: Michael Adam <obnox@samba.org>