doc: clarify dataDirHostPath must be unique for each cluster
If there are multiple clusters, `dataDirHostPath` must be unique
for each cluster. However, it's not documented yet.
In addition, `ceph-teardown.md` was also updated because
the original document assumes that `dataDirHostPath` is
`/var/lib/rook` and the files for a cluster `rook-ceph` is
under `/var/lib/rook/rook-ceph`.

related issue:
https://github.com/rook/rook/issues/14790

Signed-off-by: Satoru Takeuchi <satoru.takeuchi@gmail.com>